# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 拼单平台接口
  version: 0.2.0
  description: |
    T001：框架元数据与健康检查。
    T002：后台身份权限、商品图片、商品库存管理、小程序商品展示。
    金额一律整数分；数量为十进制字符串（≤3 位小数）；时间 ISO 8601 UTC；
    分页 page 从 1 起，pageSize 默认 10、上限 50；响应包络 {data, requestId}，错误 {code, message, requestId, details?}。
servers:
  - url: http://127.0.0.1:3000
    description: 本地联调
paths:
...
  /api/admin/v1/after-sales/tickets/{id}/refund: {"post":{"operationId":"requestReviewedRefund","summary":"客服提交实付全额退款申请，审批后执行（agent:manage、工单归属）","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["orderId","reason","clientRequestId"],"properties":{"orderId":{"type":"string","format":"uuid"},"reason":{"type":"string","minLength":1,"maxLength":1000},"clientRequestId":{"type":"string","minLength":1,"maxLength":64}}}}}},"responses":{"202":{"description":"待审核申请已保存","content":{"application/json":{"schema":{"type":"object","required":["data","requestId"],"properties":{"data":{"type":"object","required":["request"],"properties":{"request":{"$ref":"#/components/schemas/AfterSalesActionRequest"}}},"requestId":{"type":"string"}}}}}},"400":{"description":"请求或权限错误"},"401":{"description":"请求或权限错误"},"403":{"description":"请求或权限错误"},"404":{"description":"请求或权限错误"},"409":{"description":"请求或权限错误"}}}}
...
  /api/mini/v1/conversations/{id}/images:
    post:
      operationId: uploadMiniChatImage
      summary: "私有会话图片，5MiB以内，60至6000像素，校验实际类型"
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      requestBody: {"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["file"],"properties":{"file":{"type":"string","format":"binary"}}}}}}
      responses:
        '201': {"description":"私有会话图片，5MiB以内，60至6000像素，校验实际类型","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/CsImage"},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/admin/v1/cs/conversations/{id}/images:
    post:
      operationId: uploadCsChatImage
      summary: "私有会话图片，5MiB以内，60至6000像素，校验实际类型"
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      requestBody: {"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["file"],"properties":{"file":{"type":"string","format":"binary"}}}}}}
      responses:
        '201': {"description":"私有会话图片，5MiB以内，60至6000像素，校验实际类型","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/CsImage"},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/mini/v1/chat-images/{id}:
    get:
      operationId: readMiniChatImage
      summary: 本人或原负责客服下载私有原图，无公开访问
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      responses:
        '200': { description: 私有原图，Cache-Control private no-store，nosniff, content: { 'image/png': { schema: { type: string, format: binary } }, 'image/jpeg': { schema: { type: string, format: binary } }, 'image/webp': { schema: { type: string, format: binary } } } }
        '404': { $ref: '#/components/responses/Error404' }
        '401': { $ref: '#/components/responses/Error401' }
  /api/admin/v1/cs/images/{id}:
    get:
      operationId: readCsChatImage
      summary: 本人或原负责客服下载私有原图，无公开访问
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      responses:
        '200': { description: 私有原图，Cache-Control private no-store，nosniff, content: { 'image/png': { schema: { type: string, format: binary } }, 'image/jpeg': { schema: { type: string, format: binary } }, 'image/webp': { schema: { type: string, format: binary } } } }
        '404': { $ref: '#/components/responses/Error404' }
        '401': { $ref: '#/components/responses/Error401' }
...
  /api/admin/v1/after-sales/tickets/{id}/reshipment: {"post":{"operationId":"requestReviewedReshipment","summary":"客服提交补发申请，审批后执行（agent:manage、工单归属）","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["orderId","reason","clientRequestId","fulfillmentId","quantityGrams","company","trackingNo"],"properties":{"orderId":{"type":"string","format":"uuid"},"reason":{"type":"string","minLength":1,"maxLength":200},"clientRequestId":{"type":"string","minLength":1,"maxLength":64},"fulfillmentId":{"type":"string","format":"uuid"},"quantityGrams":{"type":"integer","minimum":1,"maximum":2147483647,"description":"最小履约单位：重量为克，计数为商品单位"},"company":{"type":"string","minLength":1,"maxLength":30},"trackingNo":{"type":"string","minLength":1,"maxLength":64}}}}}},"responses":{"202":{"description":"待审核申请已保存","content":{"application/json":{"schema":{"type":"object","required":["data","requestId"],"properties":{"data":{"type":"object","required":["request"],"properties":{"request":{"$ref":"#/components/schemas/AfterSalesActionRequest"}}},"requestId":{"type":"string"}}}}}},"400":{"description":"请求或权限错误"},"401":{"description":"请求或权限错误"},"403":{"description":"请求或权限错误"},"404":{"description":"请求或权限错误"},"409":{"description":"请求或权限错误"}}}}
  /api/admin/v1/after-sales/tickets/{id}/requests: {"get":{"operationId":"listTicketApprovals","summary":"工单归属范围内查看申请","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","required":["data","requestId"],"properties":{"data":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/AfterSalesActionRequest"}}}},"requestId":{"type":"string"}}}}}},"400":{"description":"请求或权限错误"},"401":{"description":"请求或权限错误"},"403":{"description":"请求或权限错误"},"404":{"description":"请求或权限错误"},"409":{"description":"请求或权限错误"}}}}
  /api/admin/v1/after-sales/tickets/{id}/fulfillment: {"get":{"operationId":"getTicketFulfillmentSummary","summary":"工单归属范围内履约摘要，不返回地址","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","required":["data","requestId"],"properties":{"data":{"type":"object","properties":{"fulfillment":{"type":"object","nullable":true,"properties":{"id":{"type":"string","format":"uuid"},"unit":{"type":"string"},"status":{"type":"string"},"allocatedQuantityGrams":{"type":"integer"},"shippedQuantityGrams":{"type":"integer"},"refundHeld":{"type":"boolean"}}}}},"requestId":{"type":"string"}}}}}},"400":{"description":"请求或权限错误"},"401":{"description":"请求或权限错误"},"403":{"description":"请求或权限错误"},"404":{"description":"请求或权限错误"},"409":{"description":"请求或权限错误"}}}}
  /api/admin/v1/after-sales/requests/{id}/review: {"post":{"operationId":"reviewAfterSalesRequest","summary":"仅超级管理员审核，重复同决定同原因返回原结果；执行/action/审计同事务","security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["decision","reason"],"properties":{"decision":{"type":"string","enum":["approve","reject"]},"reason":{"type":"string","minLength":1,"maxLength":1000}}}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","required":["data","requestId"],"properties":{"data":{"type":"object","required":["request"],"properties":{"request":{"$ref":"#/components/schemas/AfterSalesActionRequest"}}},"requestId":{"type":"string"}}}}}},"400":{"description":"请求或权限错误"},"401":{"description":"请求或权限错误"},"403":{"description":"请求或权限错误"},"404":{"description":"请求或权限错误"},"409":{"description":"请求或权限错误"}}}}
  /api/admin/v1/fulfillment/blocks/{groupId}/orders/{orderId}/refund-request: {"post":{"operationId":"requestAnomalyRefund","summary":"真实零分配异常订单创建工单与待审核全额退款申请","security":[{"bearerAuth":[]}],"parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["reason","clientRequestId"],"properties":{"reason":{"type":"string","minLength":1,"maxLength":1000},"clientRequestId":{"type":"string","minLength":1,"maxLength":64}}}}}},"responses":{"202":{"description":"待审核申请已保存","content":{"application/json":{"schema":{"type":"object","required":["data","requestId"],"properties":{"data":{"type":"object","required":["request"],"properties":{"request":{"$ref":"#/components/schemas/AfterSalesActionRequest"}}},"requestId":{"type":"string"}}}}}},"400":{"description":"请求或权限错误"},"401":{"description":"请求或权限错误"},"403":{"description":"请求或权限错误"},"404":{"description":"请求或权限错误"},"409":{"description":"请求或权限错误"}}}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 管理员会话 token（admin 域）
    miniUserAuth:
      type: http
      scheme: bearer
      description: 小程序用户会话 token（user 域，与 admin 不可互用）
  responses:
...
    Error400:
      description: 参数校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error401:
      description: 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error404:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    AfterSalesActionRequest: {"type":"object","required":["id","ticketId","orderId","kind","status","reason","payload","createdAt"],"properties":{"id":{"type":"string","format":"uuid"},"ticketId":{"type":"string","format":"uuid"},"orderId":{"type":"string","format":"uuid"},"kind":{"type":"string","enum":["refund","reshipment"]},"status":{"type":"string","enum":["pending","executed","rejected"]},"reason":{"type":"string","minLength":1,"maxLength":1000},"payload":{"type":"object"},"amountFen":{"type":"integer","nullable":true,"description":"退款为实付全额分，补发为null"},"reviewReason":{"type":"string","nullable":true},"resultId":{"type":"string","format":"uuid","nullable":true},"createdAt":{"type":"string","format":"date-time"},"reviewedAt":{"type":"string","format":"date-time","nullable":true},"requesterId":{"type":"string","format":"uuid"},"reviewerId":{"type":"string","format":"uuid","nullable":true}},"description":"用户工单详情 requests 不包含 requesterId/reviewerId；executed 表示退款请求或补发包裹已创建，不能表示退款到账"}
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items: { type: string }
...
    CsImage: {"type":"object","properties":{"id":{"type":"string","format":"uuid"},"width":{"type":"integer"},"height":{"type":"integer"}}}
...